backend/requirements.txt
Flask
psycopg2-binary
werkzeug
flask-cors